store/tooling: splitting out auth.go code before flexibility changes
